Add app launch utility
